+++ packages/ui/src/components/help/index.ts
@@ -1,0 +1,20 @@
+export {
+  type HelpArticleGroup,
+  HelpArticleIndex,
+  type HelpArticleIndexProps,
+} from "./HelpArticleIndex";
+export { HelpArticleView, type HelpArticleViewProps } from "./HelpArticleView";
+export { HelpButton, type HelpButtonProps } from "./HelpButton";
+export {
+  type HelpArticle,
+  type HelpContextValue,
+  HelpProvider,
+  type HelpProviderProps,
+  type HelpSearchResult,
+  useHelp,
+} from "./HelpProvider";
+export {
+  HelpSearchResults,
+  type HelpSearchResultsProps,
+} from "./HelpSearchResults";
+export { HelpTooltip, type HelpTooltipProps } from "./HelpTooltip";